// ClienteTryCode/clientApp/src/app/components/model/Persona.ts
export interface Persona{
    idPersona?: number;
    nombrePersona?: string;
    apellidoPersona?: string;
    edadPersona?: number;
    alturaPersona?: number;
    pesoPersona?: number;
    generoPersona?: number;
    fechaNacimientoPersona?: string;
    idPlaneta?: number;
    countViewsPersona? : number;
}